refactor(direct-control): extract production choice wrapper
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only requestCiv7ProductionChoice orchestration, the private production-choice command builder, production argument validation, read-only status payload reads, and bounded post-send polling into packages/civ7-direct-control/src/play/operations/production-choice.ts while keeping the public facade export in index.ts. The facade injects approval, ComponentID validation, App UI execution, city-operation validation, payload parsing, and serialization.

Preserves approval-first BUILD request behavior, validator-first sends, invalid-prevalidation no-send status payloads, production postcondition classification, and package/CLI production proof. Generic operation wrapper composition, telemetry, AI ingestion, semantic CLI projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- production-choice.test.ts
- bun run --cwd packages/cli test -- game.play.production.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- ownership and no-cycle rg scans

Local package/source relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
