+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -229,6 +229,15 @@ runtime/direct-control claims.
         panel closeout, validation, and postcondition coverage.
   - [x] 4.6.12 Extract unit-target action source and wrapper owner while
         keeping the public facade export surface in `index.ts`.
+  - [x] 4.6.12 Extract production-choice wrapper owner while keeping the public
+        facade export surface in `index.ts`, preserving approval-first BUILD
+        request orchestration, cityId and production-args validation,
+        validator-first behavior, read-only status payload for invalid
+        pre-validation, bounded post-send polling, production postcondition
+        classification, and package/CLI production proof, and leaving generic
+        operation wrappers, telemetry, AI ingestion, semantic CLI projection,
+        Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance
+        pending.
 - [x] 4.7 Extract settlement/tactical/progression read atoms.
   - [x] 4.7.1 Extract settlement recommendation embedded source owner while
         keeping the public wrapper in the facade.
